@jkanywhere 2021-07-06
Port of https://github.com/dgrijalva/jwt-go/pull/455.