@jkanywhere 2021-07-10
> This is a nice optimisation.

Thanks.

> Could you possibly add benchmark output to the PR comment, to demonstrate
> the savings made by this change?

Done.