@amnonbc 2021-07-10
> > Could you possibly add benchmark output to the PR comment, to demonstrate
> > the savings made by this change?
> 
> Done.

Cool...
Less allocations everywhere...
Some benchmarks actually get a bit slower....
Is this measurement noise?
I'll look at this when I get time. 
But the change is good. I say let's land this!